# Refactor combine_scorefiles