--- a/pgscatalog_utils/scorefile/config.py
+++ b/pgscatalog_utils/scorefile/config.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Config:
+    drop_missing: bool
+    liftover: bool
+    chain_dir: str
+    min_lift: float