--- a/pgscatalog_utils/scorefile/config.py
+++ b/pgscatalog_utils/scorefile/config.py
@@ -3,7 +3,9 @@ from dataclasses import dataclass
 
 @dataclass
 class Config:
+    threads: int
     drop_missing: bool
     liftover: bool
     chain_dir: str
     min_lift: float
+    batch_size: int