--- a/pgscatalog_utils/scorefile/effecttype.py
+++ b/pgscatalog_utils/scorefile/effecttype.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class EffectType(Enum):
+    RECESSIVE = "recessive"
+    DOMINANT = "dominant"
+    ADDITIVE = "additive"
+
+    def __str__(self):
+        return str(self.value)