--- a/pgscatalog_utils/scorefile/effecttype.py
+++ b/pgscatalog_utils/scorefile/effecttype.py
@@ -8,3 +8,7 @@ class EffectType(Enum):
 
     def __str__(self):
         return str(self.value)
+
+    def __repr__(self):
+        # pasting __repr__ output should be sufficient to construct the class
+        return f"{type(self).__name__}.{self.name}"