… .pre-commit-config.yaml
… pgscatalog_utils/__init__.py
… pgscatalog_utils/scorefile/effect_type.py
… pgscatalog_utils/scorefile/effect_weight.py
… pgscatalog_utils/scorefile/genome_build.py
… pgscatalog_utils/scorefile/harmonised.py
… pgscatalog_utils/scorefile/header.py
… pgscatalog_utils/scorefile/liftover.py
… pgscatalog_utils/scorefile/qc.py
… pgscatalog_utils/scorefile/read.py
… pgscatalog_utils/scorefile/scorevariant.py
… pgscatalog_utils/scorefile/scoringfile.py
… pgscatalog_utils/scorefile/scoringfileheader.py
… pgscatalog_utils/scorefile/write.py
… poetry.lock
… pyproject.toml
… tests/data/combine/PGS001229_22.txt
… tests/data/combine/scorefile.txt
… tests/data/combine/scorefile_dominant_and_recessive.txt
… tests/test_combine.py
… tests/test_liftover.py