- Old implementation used pandas to load big dataframes, do QC, and append to a gzipped file
- Tried multiprocessing #61 but the performance benefits weren't great (not CPU bound)
- New implementation sets up a data processing pipeline with a pile of generator functions
  - The end result streams variants to an output file, so memory usage will be low 
- Also fixes some bugs that cropped up:
  - pandas introduced effect weight precision errors
  - some variants lost information in weird conditions
 - `--threads` affects compression 
 
Benchmarks:

```
(new) combine_scorefiles -s scores/100/* --threads 2 -o newtest.txt.gz -t GRCh38  390.08s user 13.19s system 99% cpu 6:43.62 total 
(new, sqlite) combine_scorefiles -s scores/100/* -v -o newtest.sqlite -t GRCh38  302.38s user 11.39s system 99% cpu 5:16.23 total 
(old) combine_scorefiles -s scores/100/* -v -o oldtest.txt.gz -t GRCh38  485.12s user 34.51s system 100% cpu 8:38.01 total 
```

Future work

- Profiling shows the process is I/O bound, so try `asyncio` 
- Update match_variants to accept sqlite input