@nebfield 2023-11-07
correlation of effect weights is 1 between old and new despite the floating point error in the old version